Add arithmetic/comparison modules
