+++ py-polars/src/series.rs
@@ -1,4 +1,6 @@
-mod constructors;
+mod arithmetic;
+mod comparison;
+mod construction;
 
 use polars_algo::{cut, hist, qcut};
 use polars_core::prelude::QuantileInterpolOptions;
@@ -278,30 +280,6 @@ impl PySeries {
         } else {
             Err(PyRuntimeError::new_err("Expected a boolean mask"))
         }
-    }
-
-    pub fn add(&self, other: &PySeries) -> PyResult<Self> {
-        let out = self
-            .series
-            .try_add(&other.series)
-            .map_err(PyPolarsErr::from)?;
-        Ok(out.into())
-    }
-
-    pub fn sub(&self, other: &PySeries) -> Self {
-        (&self.series - &other.series).into()
-    }
-
-    pub fn mul(&self, other: &PySeries) -> Self {
-        (&self.series * &other.series).into()
-    }
-
-    pub fn div(&self, other: &PySeries) -> Self {
-        (&self.series / &other.series).into()
-    }
-
-    pub fn rem(&self, other: &PySeries) -> Self {
-        (&self.series % &other.series).into()
     }
 
     pub fn sort(&mut self, descending: bool) -> Self {
@@ -1072,280 +1050,6 @@ impl_get!(get_str, utf8, &str);
 impl_get!(get_date, date, i32);
 impl_get!(get_datetime, datetime, i64);
 impl_get!(get_duration, duration, i64);
-
-macro_rules! impl_arithmetic {
-    ($name:ident, $type:ty, $operand:tt) => {
-        #[pymethods]
-        impl PySeries {
-            pub fn $name(&self, other: $type) -> PyResult<Self> {
-                Ok((&self.series $operand other).into())
-            }
-        }
-    };
-}
-
-impl_arithmetic!(add_u8, u8, +);
-impl_arithmetic!(add_u16, u16, +);
-impl_arithmetic!(add_u32, u32, +);
-impl_arithmetic!(add_u64, u64, +);
-impl_arithmetic!(add_i8, i8, +);
-impl_arithmetic!(add_i16, i16, +);
-impl_arithmetic!(add_i32, i32, +);
-impl_arithmetic!(add_i64, i64, +);
-impl_arithmetic!(add_datetime, i64, +);
-impl_arithmetic!(add_duration, i64, +);
-impl_arithmetic!(add_f32, f32, +);
-impl_arithmetic!(add_f64, f64, +);
-impl_arithmetic!(sub_u8, u8, -);
-impl_arithmetic!(sub_u16, u16, -);
-impl_arithmetic!(sub_u32, u32, -);
-impl_arithmetic!(sub_u64, u64, -);
-impl_arithmetic!(sub_i8, i8, -);
-impl_arithmetic!(sub_i16, i16, -);
-impl_arithmetic!(sub_i32, i32, -);
-impl_arithmetic!(sub_i64, i64, -);
-impl_arithmetic!(sub_datetime, i64, -);
-impl_arithmetic!(sub_duration, i64, -);
-impl_arithmetic!(sub_f32, f32, -);
-impl_arithmetic!(sub_f64, f64, -);
-impl_arithmetic!(div_u8, u8, /);
-impl_arithmetic!(div_u16, u16, /);
-impl_arithmetic!(div_u32, u32, /);
-impl_arithmetic!(div_u64, u64, /);
-impl_arithmetic!(div_i8, i8, /);
-impl_arithmetic!(div_i16, i16, /);
-impl_arithmetic!(div_i32, i32, /);
-impl_arithmetic!(div_i64, i64, /);
-impl_arithmetic!(div_f32, f32, /);
-impl_arithmetic!(div_f64, f64, /);
-impl_arithmetic!(mul_u8, u8, *);
-impl_arithmetic!(mul_u16, u16, *);
-impl_arithmetic!(mul_u32, u32, *);
-impl_arithmetic!(mul_u64, u64, *);
-impl_arithmetic!(mul_i8, i8, *);
-impl_arithmetic!(mul_i16, i16, *);
-impl_arithmetic!(mul_i32, i32, *);
-impl_arithmetic!(mul_i64, i64, *);
-impl_arithmetic!(mul_f32, f32, *);
-impl_arithmetic!(mul_f64, f64, *);
-impl_arithmetic!(rem_u8, u8, %);
-impl_arithmetic!(rem_u16, u16, %);
-impl_arithmetic!(rem_u32, u32, %);
-impl_arithmetic!(rem_u64, u64, %);
-impl_arithmetic!(rem_i8, i8, %);
-impl_arithmetic!(rem_i16, i16, %);
-impl_arithmetic!(rem_i32, i32, %);
-impl_arithmetic!(rem_i64, i64, %);
-impl_arithmetic!(rem_f32, f32, %);
-impl_arithmetic!(rem_f64, f64, %);
-
-macro_rules! impl_rhs_arithmetic {
-    ($name:ident, $type:ty, $operand:ident) => {
-        #[pymethods]
-        impl PySeries {
-            pub fn $name(&self, other: $type) -> PyResult<Self> {
-                Ok(other.$operand(&self.series).into())
-            }
-        }
-    };
-}
-
-impl_rhs_arithmetic!(add_u8_rhs, u8, add);
-impl_rhs_arithmetic!(add_u16_rhs, u16, add);
-impl_rhs_arithmetic!(add_u32_rhs, u32, add);
-impl_rhs_arithmetic!(add_u64_rhs, u64, add);
-impl_rhs_arithmetic!(add_i8_rhs, i8, add);
-impl_rhs_arithmetic!(add_i16_rhs, i16, add);
-impl_rhs_arithmetic!(add_i32_rhs, i32, add);
-impl_rhs_arithmetic!(add_i64_rhs, i64, add);
-impl_rhs_arithmetic!(add_f32_rhs, f32, add);
-impl_rhs_arithmetic!(add_f64_rhs, f64, add);
-impl_rhs_arithmetic!(sub_u8_rhs, u8, sub);
-impl_rhs_arithmetic!(sub_u16_rhs, u16, sub);
-impl_rhs_arithmetic!(sub_u32_rhs, u32, sub);
-impl_rhs_arithmetic!(sub_u64_rhs, u64, sub);
-impl_rhs_arithmetic!(sub_i8_rhs, i8, sub);
-impl_rhs_arithmetic!(sub_i16_rhs, i16, sub);
-impl_rhs_arithmetic!(sub_i32_rhs, i32, sub);
-impl_rhs_arithmetic!(sub_i64_rhs, i64, sub);
-impl_rhs_arithmetic!(sub_f32_rhs, f32, sub);
-impl_rhs_arithmetic!(sub_f64_rhs, f64, sub);
-impl_rhs_arithmetic!(div_u8_rhs, u8, div);
-impl_rhs_arithmetic!(div_u16_rhs, u16, div);
-impl_rhs_arithmetic!(div_u32_rhs, u32, div);
-impl_rhs_arithmetic!(div_u64_rhs, u64, div);
-impl_rhs_arithmetic!(div_i8_rhs, i8, div);
-impl_rhs_arithmetic!(div_i16_rhs, i16, div);
-impl_rhs_arithmetic!(div_i32_rhs, i32, div);
-impl_rhs_arithmetic!(div_i64_rhs, i64, div);
-impl_rhs_arithmetic!(div_f32_rhs, f32, div);
-impl_rhs_arithmetic!(div_f64_rhs, f64, div);
-impl_rhs_arithmetic!(mul_u8_rhs, u8, mul);
-impl_rhs_arithmetic!(mul_u16_rhs, u16, mul);
-impl_rhs_arithmetic!(mul_u32_rhs, u32, mul);
-impl_rhs_arithmetic!(mul_u64_rhs, u64, mul);
-impl_rhs_arithmetic!(mul_i8_rhs, i8, mul);
-impl_rhs_arithmetic!(mul_i16_rhs, i16, mul);
-impl_rhs_arithmetic!(mul_i32_rhs, i32, mul);
-impl_rhs_arithmetic!(mul_i64_rhs, i64, mul);
-impl_rhs_arithmetic!(mul_f32_rhs, f32, mul);
-impl_rhs_arithmetic!(mul_f64_rhs, f64, mul);
-impl_rhs_arithmetic!(rem_u8_rhs, u8, rem);
-impl_rhs_arithmetic!(rem_u16_rhs, u16, rem);
-impl_rhs_arithmetic!(rem_u32_rhs, u32, rem);
-impl_rhs_arithmetic!(rem_u64_rhs, u64, rem);
-impl_rhs_arithmetic!(rem_i8_rhs, i8, rem);
-impl_rhs_arithmetic!(rem_i16_rhs, i16, rem);
-impl_rhs_arithmetic!(rem_i32_rhs, i32, rem);
-impl_rhs_arithmetic!(rem_i64_rhs, i64, rem);
-impl_rhs_arithmetic!(rem_f32_rhs, f32, rem);
-impl_rhs_arithmetic!(rem_f64_rhs, f64, rem);
-
-macro_rules! impl_eq_num {
-    ($name:ident, $type:ty) => {
-        #[pymethods]
-        impl PySeries {
-            pub fn $name(&self, rhs: $type) -> PyResult<Self> {
-                let s = self.series.equal(rhs).map_err(PyPolarsErr::from)?;
-                Ok(s.into_series().into())
-            }
-        }
-    };
-}
-
-impl_eq_num!(eq_u8, u8);
-impl_eq_num!(eq_u16, u16);
-impl_eq_num!(eq_u32, u32);
-impl_eq_num!(eq_u64, u64);
-impl_eq_num!(eq_i8, i8);
-impl_eq_num!(eq_i16, i16);
-impl_eq_num!(eq_i32, i32);
-impl_eq_num!(eq_i64, i64);
-impl_eq_num!(eq_f32, f32);
-impl_eq_num!(eq_f64, f64);
-impl_eq_num!(eq_str, &str);
-
-macro_rules! impl_neq_num {
-    ($name:ident, $type:ty) => {
-        #[allow(clippy::nonstandard_macro_braces)]
-        #[pymethods]
-        impl PySeries {
-            pub fn $name(&self, rhs: $type) -> PyResult<Self> {
-                let s = self.series.not_equal(rhs).map_err(PyPolarsErr::from)?;
-                Ok(s.into_series().into())
-            }
-        }
-    };
-}
-
-impl_neq_num!(neq_u8, u8);
-impl_neq_num!(neq_u16, u16);
-impl_neq_num!(neq_u32, u32);
-impl_neq_num!(neq_u64, u64);
-impl_neq_num!(neq_i8, i8);
-impl_neq_num!(neq_i16, i16);
-impl_neq_num!(neq_i32, i32);
-impl_neq_num!(neq_i64, i64);
-impl_neq_num!(neq_f32, f32);
-impl_neq_num!(neq_f64, f64);
-impl_neq_num!(neq_str, &str);
-
-macro_rules! impl_gt_num {
-    ($name:ident, $type:ty) => {
-        #[pymethods]
-        impl PySeries {
-            pub fn $name(&self, rhs: $type) -> PyResult<Self> {
-                let s = self.series.gt(rhs).map_err(PyPolarsErr::from)?;
-                Ok(s.into_series().into())
-            }
-        }
-    };
-}
-
-impl_gt_num!(gt_u8, u8);
-impl_gt_num!(gt_u16, u16);
-impl_gt_num!(gt_u32, u32);
-impl_gt_num!(gt_u64, u64);
-impl_gt_num!(gt_i8, i8);
-impl_gt_num!(gt_i16, i16);
-impl_gt_num!(gt_i32, i32);
-impl_gt_num!(gt_i64, i64);
-impl_gt_num!(gt_f32, f32);
-impl_gt_num!(gt_f64, f64);
-impl_gt_num!(gt_str, &str);
-
-macro_rules! impl_gt_eq_num {
-    ($name:ident, $type:ty) => {
-        #[pymethods]
-        impl PySeries {
-            pub fn $name(&self, rhs: $type) -> PyResult<Self> {
-                let s = self.series.gt_eq(rhs).map_err(PyPolarsErr::from)?;
-                Ok(s.into_series().into())
-            }
-        }
-    };
-}
-
-impl_gt_eq_num!(gt_eq_u8, u8);
-impl_gt_eq_num!(gt_eq_u16, u16);
-impl_gt_eq_num!(gt_eq_u32, u32);
-impl_gt_eq_num!(gt_eq_u64, u64);
-impl_gt_eq_num!(gt_eq_i8, i8);
-impl_gt_eq_num!(gt_eq_i16, i16);
-impl_gt_eq_num!(gt_eq_i32, i32);
-impl_gt_eq_num!(gt_eq_i64, i64);
-impl_gt_eq_num!(gt_eq_f32, f32);
-impl_gt_eq_num!(gt_eq_f64, f64);
-impl_gt_eq_num!(gt_eq_str, &str);
-
-macro_rules! impl_lt_num {
-    ($name:ident, $type:ty) => {
-        #[allow(clippy::nonstandard_macro_braces)]
-        #[pymethods]
-        impl PySeries {
-            pub fn $name(&self, rhs: $type) -> PyResult<Self> {
-                let s = self.series.lt(rhs).map_err(PyPolarsErr::from)?;
-                Ok(s.into_series().into())
-            }
-        }
-    };
-}
-
-impl_lt_num!(lt_u8, u8);
-impl_lt_num!(lt_u16, u16);
-impl_lt_num!(lt_u32, u32);
-impl_lt_num!(lt_u64, u64);
-impl_lt_num!(lt_i8, i8);
-impl_lt_num!(lt_i16, i16);
-impl_lt_num!(lt_i32, i32);
-impl_lt_num!(lt_i64, i64);
-impl_lt_num!(lt_f32, f32);
-impl_lt_num!(lt_f64, f64);
-impl_lt_num!(lt_str, &str);
-
-macro_rules! impl_lt_eq_num {
-    ($name:ident, $type:ty) => {
-        #[pymethods]
-        impl PySeries {
-            pub fn $name(&self, rhs: $type) -> PyResult<Self> {
-                let s = self.series.lt_eq(rhs).map_err(PyPolarsErr::from)?;
-                Ok(s.into_series().into())
-            }
-        }
-    };
-}
-
-impl_lt_eq_num!(lt_eq_u8, u8);
-impl_lt_eq_num!(lt_eq_u16, u16);
-impl_lt_eq_num!(lt_eq_u32, u32);
-impl_lt_eq_num!(lt_eq_u64, u64);
-impl_lt_eq_num!(lt_eq_i8, i8);
-impl_lt_eq_num!(lt_eq_i16, i16);
-impl_lt_eq_num!(lt_eq_i32, i32);
-impl_lt_eq_num!(lt_eq_i64, i64);
-impl_lt_eq_num!(lt_eq_f32, f32);
-impl_lt_eq_num!(lt_eq_f64, f64);
-impl_lt_eq_num!(lt_eq_str, &str);
 
 #[cfg(test)]
 mod test {
